Yajem-1 User equipment -> vorbereitende Maßnahmen. - user Helper yajemUser
