+++ administrator/components/com_yajem/views/events/view.html.php
@@ -54,7 +54,7 @@ class YajemViewEvents extends HtmlView
 	{
 		$this->items      = $this->get('Items');
 		for ($i=0; $i<count($this->items);$i++){
-			$this->items[$i]->organizer = YajemHelperAdmin::getUser($this->items[$i]->organizerId);
+			$this->items[$i]->organizer = YajemUserHelperAdmin::getUser($this->items[$i]->organizerId);
 		}
 		$this->pagination = $this->get('Pagination');
 		$this->state      = $this->get('State');
